# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API documentation for E-commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Ecommerce/public",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/Translations/category/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get category translation languages",
                "description": "Retrieve available and unavailable translation languages for a specific category.",
                "operationId": "4cea995e754dde11aa11b9520135c815",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Languages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "available_langs": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "en",
                                                "ar"
                                            ]
                                        },
                                        "available_langs_count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "unavailable_langs": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "fr",
                                                "de"
                                            ]
                                        },
                                        "unavailable_langs_count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/Translations/category/add/{id}": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Add translation to category",
                "description": "Add a new translation for a specific category.",
                "operationId": "7b9014eb193e3fcb1fb7ae5a0ec992b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "lang"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Electronics",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Electrónica y dispositivos modernos"
                                    },
                                    "lang": {
                                        "$ref": "#/components/schemas/LanguagesEnum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Translation added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Language already exists for this category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This language already exists for this category."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The name field is required."
                                                    ]
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The description field is required."
                                                    ]
                                                },
                                                "lang": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The selected lang is invalid."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/Translations/product/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get product translation languages",
                "description": "Retrieve available and unavailable translation languages for a specific product owned by the authenticated user.",
                "operationId": "9a9dd70d36a328c2056f4407be5a6e53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Languages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "available_langs": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "en",
                                                "ar"
                                            ]
                                        },
                                        "available_langs_count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "unavailable_langs": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "fr",
                                                "de"
                                            ]
                                        },
                                        "unavailable_langs_count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Product not found or unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/Translations/product/add/{id}": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Add translation to product",
                "description": "Add a new translation for a specific product owned by the authenticated user.",
                "operationId": "1d36448261d036b9856ff2524f152042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "lang"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Portátil de alto rendimiento"
                                    },
                                    "lang": {
                                        "$ref": "#/components/schemas/LanguagesEnum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Translation added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Language already exists for this product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This language already exists for this product."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Product not found or unauthorized"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The name field is required."
                                                    ]
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The description field is required."
                                                    ]
                                                },
                                                "lang": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The selected lang is invalid."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LanguagesEnum": {
                "type": "string",
                "enum": [
                    "ar",
                    "en",
                    "sp",
                    "ur"
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}